+++ l10n/fr.js
@@ -1,6 +1,28 @@
 OC.L10N.register(
     "gpxedit",
     {
+    "Draw a track" : "Dessiner une trace",
+    "Add a waypoint" : "Ajouter une étape",
+    "Edit" : "Éditer",
+    "Nothing to edit" : "Rien à éditer",
+    "Delete" : "Effacer",
+    "Nothing to delete" : "Rien à effacer",
+    "Cancel" : "Annuler",
+    "Discard all changes" : "Annuler tous les changements",
+    "Validate changes" : "Valider les changements",
+    "Drag to move elements,<br/>click to remove a point" : "Déplacer par glisser-déposer,<br/>cliquer sur un point pour le supprimer",
+    "Click cancel to undo changes": "Cliquer sur annuler pour défaire les changements",
+    "Click on an element to delete it" : "Cliquer sur un élément pour l'effacer",
+    "Click map to add waypoint" : "Cliquer sur la carte pour ajouter une étape",
+    "Click to start drawing track" : "Cliquer pour commencer à dessiner une trace",
+    "Click to continue drawing track" : "Cliquer pour continuer à dessiner une trace",
+    "Click last point to finish track" : "Cliquer sur le dernier point pour terminer la trace",
+    "Cancel drawing" : "Annuler le dessin",
+    "Finish" : "Terminer",
+    "Finish drawing" : "Terminer le dessin",
+    "Delete last point" : "Supprimer le dernier point",
+    "Delete last point drawn" : "Supprimer le dernier point dessiné"
+
     },
 "nplurals=2; plural=(n > 1);");